+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next 
+
+- Added `:extensions` for custom file type extensions
+
 # 0.4.3 - Mar 21, 2024
 
 - Do not report self-reference as a cycle #6